refactor(index): вынести extract_signature в index/struct_diff с ре-экспортом (PRI-158)
